fix: move mobile navbar outside header for proper iOS Safari fixed positioning
- Create separate MobileNavbar component at root layout level
- Remove mobile nav from Navbar component to avoid header positioning constraints
- Use z-[9999] to ensure navbar stays on top of all content
- Use inline style for padding-bottom to properly apply safe-area-inset
- Mobile nav now truly fixed at viewport bottom on all devices including iOS
